% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift.R
\docType{methods}
\name{Redshift}
\alias{Redshift}
\alias{dbConnect,RedshiftDriver-method}
\title{Redshift DBI wrapper}
\usage{
Redshift()

\S4method{dbConnect}{RedshiftDriver}(drv, host, dbname, user, password, ...)
}
\arguments{
\item{drv}{An object created by \code{Redshift()}}

\item{host}{the host}

\item{dbname}{S3 bucket where results will be saved to}

\item{user}{Redshift schema to use}

\item{password}{pw}

\item{...}{Other options}
}
\description{
Redshift DBI wrapper

Authentication credentials are read from the DefaultAWSCredentialsProviderChain, which includes the .aws folder and
environment variables.
}
\examples{
\dontrun{
require(DBI)
con <- dbConnect(AWR.Redshift::Redshift(), region='us-west-2', 
                 s3_staging_dir='s3://nfultz-Redshift-staging', 
                 schema_name='default')
dbListTables(con)
dbGetQuery(con, "Select count(*) from sampledb.elb_logs")
}
}
\seealso{
\href{http://docs.aws.amazon.com/Redshift/latest/ug/connect-with-jdbc.html#jdbc-options}{Redshift Manual} for more connections options.
}
