% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redshift.R
\docType{class}
\name{RedshiftDriver-class}
\alias{RedshiftDriver-class}
\alias{RedshiftDriver}
\title{Redshift driver class.}
\usage{
\S4method{initialize}{RedshiftDriver}(.Object, ...)
}
\description{
Redshift driver class.

Constructor of RedshiftDriver
}
\keyword{internal}
